wheel
This code creates an array wheelOptions containing different possible options for the wheel of fortune. Then, it defines the spinWheel() function, which selects a random element from the wheelOptions array and returns its value. In this example, the spin result is logged to the console using console.log(), but it can be easily modified to display on a web page or in another format as needed.